Add object rest spread support
Uses acorn-node which is Acorn + some plugins and default options to
match recent Node.js.

This also changes the `opts` parameter to a simple pass thru, so that we
only need to update acorn-node when a new `ecmaVersion` is supported in
Node.

If this is merged we don't need #19 because you will already be able to
pass the `sourceType`.